rename mizu_validator to mizu_node_worker
