+++ mizu_node_worker/main.py
@@ -5,7 +5,7 @@ from fastapi import FastAPI
 from redis import Redis
 from rq import Queue, Worker
 
-from mizu_validator.worker import job_worker, WorkerJob
+from mizu_node_worker.worker import job_worker, WorkerJob
 
 # HTTP server entry point
 app = FastAPI()
@@ -25,12 +25,12 @@ def do_classify(job: WorkerJob):
 
 
 def start_dev():
-    uvicorn.run("mizu_validator.main:app", host="0.0.0.0", port=8001, reload=True)
+    uvicorn.run("mizu_node_worker.main:app", host="0.0.0.0", port=8001, reload=True)
 
 
 # the number of workers is defined by $WEB_CONCURRENCY env as default
 def start():
-    uvicorn.run("mizu_validator.main:app", host="0.0.0.0", port=8001)
+    uvicorn.run("mizu_node_worker.main:app", host="0.0.0.0", port=8001)
 
 
 def start_rq_worker():
